Move triggers.generated to the bottom again
This speeds up extension installation.
