+++ extension/sql/load-order.txt
@@ -11,7 +11,6 @@ counter_agg.generated.sql
 utilities.generated.sql
 stats_agg.generated.sql
 time_series_pipeline.generated.sql
-triggers.generated.sql
 time_series_pipeline_delta.generated.sql
 time_series_pipeline_fill_holes.generated.sql
 time_series_pipeline_map.generated.sql
@@ -23,3 +22,4 @@ time_series_pipeline_aggregation.generated.sql
 schema_test.generated.sql
 serialization_collations.generated.sql
 serialization_types.generated.sql
+zz_triggers.generated.sql
